backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Redis / Celery
redis>=5.0.0
celery[redis]>=5.3.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.0

# File handling
aiofiles>=23.2.0
python-magic>=0.4.27

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

